+++ src/controllers/rabbitmq_controller/proof/helper_invariants/mod.rs
@@ -1,9 +1,7 @@
 // Copyright 2022 VMware, Inc.
 // SPDX-License-Identifier: MIT
-pub mod owner_ref;
 pub mod predicate;
 pub mod proof;
 
-pub use owner_ref::*;
 pub use predicate::*;
 pub use proof::*;
